social-media-command-center/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
python-dotenv==1.0.1
pydantic==2.9.2

# LangChain
langchain==0.3.1
langchain-anthropic==0.3.0
langchain-community==0.3.1

# Embeddings (free, local)
sentence-transformers==3.0.1

# Pinecone
pinecone-client==5.0.1

# Redis for task queue
redis==5.1.0

# HTTP
httpx==0.27.2
aiohttp==3.10.6

# Utilities
python-multipart==0.0.12
jinja2==3.1.4

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
